lib/string: change parameter types to match C standard
The third parameter of strncpy and strncmp is supposed to be a size_t
and not a signed int.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/83222
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/lib/string.c
@@ -71,7 +71,7 @@ char *strrchr(const char *s, int c)
 	return p;
 }
 
-char *strncpy(char *to, const char *from, int count)
+char *strncpy(char *to, const char *from, size_t count)
 {
 	char *ret = to;
 	char data;
@@ -113,7 +113,7 @@ int strcmp(const char *s1, const char *s2)
 	return r;
 }
 
-int strncmp(const char *s1, const char *s2, int maxlen)
+int strncmp(const char *s1, const char *s2, size_t maxlen)
 {
 	int i;
